Refactored `Module` into the `rusty-junctions-utils` crate
Refactored the `Module` struct into the `rusty-junctions-utils` crate,
this allows it to be very easily used across the different macros.

I intend on doing the same for the generic types parameters, and all
of the other constructs that are common across the macros.
