+++ rusty-junctions-macro/rusty-junctions-partial-pattern-derive/Cargo.toml
@@ -9,6 +9,7 @@ edition = "2021"
 syn = { version = "1.0", features = ["full"] }
 quote = "1.0"
 proc-macro2 = "1.0"
+rusty-junctions-utils = { path = "./../rusty-junctions-utils" }
 
 [lib]
 proc-macro = true
