+++ rusty-junctions-macro/rusty-junctions-partial-pattern-derive/src/lib.rs
@@ -5,6 +5,7 @@ use syn::{
     parse_macro_input, Data, DataStruct, DeriveInput, Field, Fields, FieldsNamed, GenericParam,
     Generics, TypeParam, __private::TokenStream2,
 };
+use rusty_junctions_utils::Module;
 
 mod channels;
 
@@ -260,40 +261,5 @@ impl JoinPattern {
             "SendJoinPattern" => true,
             _not_send_pattern => false,
         }
-    }
-}
-
-struct Module {
-    pub number: usize,
-}
-
-impl Module {
-    pub fn from_usize(number: usize) -> Self {
-        Self { number }
-    }
-
-    pub fn name(&self) -> String {
-        let name = match self.number {
-            0 => panic!("Invalid number of fields"),
-            1 => "unary".to_string(),
-            2 => "binary".to_string(),
-            3 => "ternary".to_string(),
-            n => format!("n{}ary", n),
-        };
-
-        name
-    }
-
-    pub fn ident(&self) -> Ident {
-        let name = self.name();
-        Ident::new(&name, Span::call_site())
-    }
-}
-
-impl std::iter::Iterator for Module {
-    type Item = Self;
-    fn next(&mut self) -> Option<<Self as std::iter::Iterator>::Item> {
-        let number = self.number + 1;
-        Some(Self { number })
     }
 }
