+++ rusty-junctions-macro/rusty-junctions-utils/Cargo.toml
@@ -6,3 +6,5 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+syn = { version = "1.0", features = ["full"] }
+proc-macro2 = "1.0"
